+++ src/dataArgs.c
@@ -294,7 +294,7 @@ char* getPrintStringFromFloat(Args* self, char* name, float val) {
 char* getPrintStringFromPtr(Args* self, char* name, void* val) {
     Args* buffs = New_strBuff();
     char* res = NULL;
-    uint64_t intVal = (long)val;
+    uint64_t intVal = (uintptr_t)val;
     char* valString = strsFormat(buffs, 32, "0x%llx", intVal);
     res = getPrintSring(self, name, valString);
     args_deinit(buffs);
